requirements.txt
streamlit>=1.40.0
chromadb>=1.0.0
google-genai>=1.0.0
rank-bm25>=0.2.2
kiwipiepy>=0.18.0
